basic page layout complete
